+++ packages/components/src/Dialog/index.ts
@@ -27,7 +27,6 @@
 export * from './Confirm'
 export * from './Dialog'
 export * from './DialogContext'
-export * from './DialogManager'
 export * from './Layout'
 export * from './Prompt'
 export * from './useDialog'